requirements.txt
python-chess>=1.999
numpy>=1.24.0
matplotlib>=3.7.0
gym>=0.26.0
stable-baselines3>=2.0.0
torch>=2.0.0
pytest>=7.4.0
jupyter>=1.0.0

